openstack-flask-app/requirements.txt
# Flask core with async support
flask[async]>=3.0.0
jinja2>=3.1.0
itsdangerous>=2.1.0
click>=8.1.0
werkzeug>=3.0.0

# OpenStack SDK (for APIs like compute, network, etc.)
openstacksdk>=1.4.0

# Async utilities
aiohttp>=3.9.0
asyncio>=3.4.3
anyio>=4.0.0

# For environment and config management
python-dotenv>=1.0.0

# Optional - for flash messages styling, templates, etc.
flask-bootstrap>=3.3.7.1
flask-wtf>=1.2.1

# Optional - for better async handling if using threading
asgiref>=3.7.2
